Updated DisplayTagHelper and EditorTagHelper.
Switched from an anonymous object to dictionary for htmlAttributes, so that only properties that have values can be added to the dictionary.
Added Id property to set the id on the html element.

Updated samples to demonstrate the merging of class, style and id supplied with editor tag.
Also demonstrates the use for view-data-htmAttributes.